web_app/requirements.txt
aiofiles==23.1.0
blinker==1.5
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
filelock==3.12.2
fsspec==2023.6.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
huggingface-hub==0.15.1
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
motor==3.1.2
numpy==1.24.3
packaging==23.1
priority==2.0.0
pyaes==1.6.1
pyasn1==0.5.0
pymongo==4.3.3
python-dotenv==1.0.0
PyYAML==6.0
quart==0.18.4
regex==2023.6.3
requests==2.31.0
rsa==4.9
safetensors==0.3.1
Telethon==1.28.5
tokenizers==0.13.3
toml==0.10.2
tqdm==4.65.0
transformers==4.30.2
typing_extensions==4.6.3
urllib3==2.0.3
Werkzeug==2.3.6
wsproto==1.2.0
